move SchemaForIPList and SchemaForValuePrefixLengthResource to common pkg
