+++ nutanix/services/clustersv2/resource_nutanix_cluster_fetch_networks_v2.go
@@ -113,7 +113,7 @@ func nodeListNetworkingDetailsSchema() *schema.Schema {
 							"cvm_ip": {
 								Type:     schema.TypeList,
 								Computed: true,
-								Elem:     SchemaForIPList(false),
+								Elem:     common.SchemaForIPList(false),
 							},
 							"uplink_list": {
 								Type:     schema.TypeList,
@@ -171,19 +171,19 @@ func nodeListSchema() *schema.Schema {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem:     SchemaForIPList(false),
+					Elem:     common.SchemaForIPList(false),
 				},
 				"hypervisor_ip": {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem:     SchemaForIPList(false),
+					Elem:     common.SchemaForIPList(false),
 				},
 				"ipmi_ip": {
 					Type:     schema.TypeList,
 					Optional: true,
 					Computed: true,
-					Elem:     SchemaForIPList(false),
+					Elem:     common.SchemaForIPList(false),
 				},
 				"digital_certificate_map_list": {
 					Type:     schema.TypeList,
